Python ML/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
numpy
scikit-learn
python-dotenv
